+++ components/core/ktx/src/main/java/com/flipperdevices/core/ktx/jre/NullKtx.kt
@@ -1,0 +1,9 @@
+package com.flipperdevices.core.ktx.jre
+
+/**
+ * @param items array objects
+ * @return is there a null object in the array
+ */
+fun isNotNull(vararg items: Any?): Boolean {
+    return items.none { it == null }
+}