+++ components/info/impl/src/main/java/com/flipperdevices/info/impl/compose/screens/ComposableFullDeviceInfoScreen.kt
@@ -7,8 +7,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.Icon
 import androidx.compose.material.Text
 import androidx.compose.material.ripple.rememberRipple
@@ -23,65 +21,41 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
-import com.flipperdevices.core.ktx.jre.titlecaseFirstCharIfItIsLowercase
 import com.flipperdevices.core.ui.res.R as DesignSystem
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 import com.flipperdevices.info.impl.R
-import com.flipperdevices.info.impl.compose.elements.ComposableInfoCardContent
+import com.flipperdevices.info.impl.compose.elements.ComposableFullInfoDevice
 import com.flipperdevices.info.impl.compose.navigation.NavGraphRoute
+import com.flipperdevices.info.impl.fragment.DeviceInfoHelper
+import com.flipperdevices.info.impl.model.DeviceStatus
 import com.flipperdevices.info.impl.viewmodel.DeviceInfoViewModel
-import com.flipperdevices.info.shared.ComposableDeviceInfoRowWithText
-import com.flipperdevices.info.shared.ComposableInfoDivider
-import com.flipperdevices.info.shared.InfoElementCard
+import com.flipperdevices.info.impl.viewmodel.DeviceStatusViewModel
 
 @Composable
 fun ComposableFullDeviceInfoScreen(
     navController: NavHostController,
-    deviceInfoViewModel: DeviceInfoViewModel = viewModel()
+    deviceInfoViewModel: DeviceInfoViewModel = viewModel(),
+    deviceStatusViewModel: DeviceStatusViewModel = viewModel()
 ) {
+    val deviceStatus by deviceStatusViewModel.getState().collectAsState()
+    val deviceInfoRequestStatus by deviceInfoViewModel.getDeviceInfoRequestStatus().collectAsState()
+    val localDeviceStatus = deviceStatus
+
+    val inProgress = if (localDeviceStatus is DeviceStatus.NoDeviceInformation) {
+        localDeviceStatus.connectInProgress
+    } else {
+        localDeviceStatus is DeviceStatus.Connected &&
+            deviceInfoRequestStatus.rpcDeviceInfoRequestInProgress
+    }
+
     val verboseDeviceInfo by deviceInfoViewModel.getVerboseDeviceInfoState().collectAsState()
-    val fieldEntriesFiltered = verboseDeviceInfo
-        .rpcInformationMap.entries.filterNot { it.key.isBlank() || it.value.isBlank() }
+    val fields: Map<String, String> = verboseDeviceInfo.rpcInformationMap
+    val deviceInfo = DeviceInfoHelper.parseFields(fields, deviceInfoViewModel::getFirmwareChannel)
 
     Column {
         ComposableFullDeviceInfoScreenBar { navController.navigate(NavGraphRoute.Info.name) }
-        Column(
-            Modifier
-                .verticalScroll(rememberScrollState())
-        ) {
-            InfoElementCard(
-                Modifier.padding(top = 14.dp),
-                R.string.info_device_info_title_card
-            ) {
-                ComposableInfoCardContent(isUnsupported = false)
-            }
-            if (fieldEntriesFiltered.isNotEmpty()) {
-                InfoElementCard(
-                    Modifier.padding(top = 14.dp, bottom = 14.dp),
-                    R.string.info_device_info_other_card
-                ) {
-                    ComposableOtherCardContent(fieldEntriesFiltered)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ComposableOtherCardContent(
-    fields: List<Map.Entry<String, String>>
-) {
-    fields.forEachIndexed { index, field ->
-        val name = field.key.split(' ', '_')
-            .map { it.titlecaseFirstCharIfItIsLowercase() }
-            .joinToString(" ")
-        val value = field.value
-
-        ComposableDeviceInfoRowWithText(text = name, inProgress = false, value = value)
-        if (index != fields.size - 1) {
-            ComposableInfoDivider()
-        }
+        ComposableFullInfoDevice(deviceInfo, inProgress)
     }
 }